Rename FabricDataProvider to FabricCodecDataProvider and remove FabricAtlasSourceProvider
